refactor(mobile): split latest switcher widget
