+++ mobile/test/features/channels/jump_to_latest_button_test.dart
@@ -1,5 +1,6 @@
 import 'package:buzz/features/channels/message_action_backdrop_state.dart';
 import 'package:buzz/features/channels/jump_to_latest_button.dart';
+import 'package:buzz/features/channels/jump_to_latest_switcher.dart';
 import 'package:buzz/shared/theme/theme.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
